feat: make it a pwa
This adds some minimal steps to make this website installable and
offline usable, along with some minor style changes so that it adapts to
smaller screens.

In order to make it installable there is a manifest file, which tells
the browser about the title of the app as well as some icons which can
be used as launching icons by the os.

Next, in order to provice offline functionality (in fact this is even
needed for the pwa to be installable), we add a ServieWorker. When the
ServieWorker is installed, it fetches all required resources and
populates a cache. It then intercepts the browsers fetch and responds
with the cached resources instead of making a network request.

Ah I could not resist to remove the canvas wrapping div, as it was not
needed - at least I did not see a reason for it. This made the screen
size adaption even simpler, although this was not strictly needed.